api/src/requirements.txt
docker
elasticsearch
elasticsearch[async]
pymongo
aiofiles
aiohttp
uvicorn
jinja2
fastapi
python-multipart
timeout-decorator
pytest
pytimeparse
pytest-cov
sqlalchemy
sqlalchemy_utils
Authlib
httpx
itsdangerous
pyjwt
passlib[bcrypt]